refactor contact component
