+++ containers/Home/Contact/RenderTypeOnPrintMode.tsx
@@ -1,0 +1,7 @@
+import { decodePhoneNumber } from './utils'
+import type { RenderTypeProps } from './interface'
+
+export const RenderTypeOnPrintMode: React.FC<RenderTypeProps> = ({ type, linkProps, content }) => {
+    const renderContent = type === 'phone' ? decodePhoneNumber(content) : content
+    return <a {...linkProps}> {renderContent} </a>
+}
